chore: Change a default port of running servers
Change port of character-maker from 8081 to random port.
Not to intersect with api-gateway's port and other ports
if they occur in the future.

Change port of characters from 8080 to random port.
Change by the same reason as previous.

Change port of api-gateway from random to 8080.
This change was being made because of api-gateway's
port which should be certain value. Unless request's
port constantly will change and testing endpoint would
be hard.