feat: Make CharacterService call the remote microservice
Add: CharacterService makes GET call to /api/characters in
order to fetch existing character's names. It uses
WebClient.Builder to do that.

Response from /api/characters is converted to
List<CharacterDto> to pass it to an if-else function.
The function checks whether CharacterDto object from POST
request has the same name as names from List<CharacterDto>.
If request's object has any match with these names it
throws RuntimeException. Otherwise, the object is saved to
a repository.

Also add Transactional annotations to class to mark it as
a transactional.